Refactor logic code to the engine in DayRangeSelectionHelper
